backend/pyproject.toml
[project]
name = "dronedata-backend"
version = "0.1.0"
description = "DroneData backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2",
    "pydantic-settings>=2",
    "sqlalchemy[asyncio]>=2",
    "asyncpg>=0.29",
    "psycopg[binary]>=3.1",
    "alembic>=1.13",
    "geoalchemy2>=0.15",
    "celery[redis]>=5.3",
    "redis>=5",
    "shapely>=2",
    "pyproj>=3.6",
    "python-multipart>=0.0.9",
    "pillow>=10",
    "pyodm>=1.5",
    "rio-cogeo>=5",
    "rasterio>=1.3",
    "sse-starlette>=2",
    "slowapi>=0.1.9",
    "boto3>=1.34",
    "numpy>=1.26",
    "opencv-python-headless>=4.9",
    "scikit-image>=0.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
